move converter code out of test suite
